Rebase ``FileDialogNative``
