+++ CollapseLauncher/XAMLs/MainApp/Pages/HomePage.xaml.cs
@@ -1,6 +1,7 @@
 using CollapseLauncher.Dialogs;
 using CollapseLauncher.Interfaces;
 using CollapseLauncher.Statics;
+using CollapseLauncher.FileDialogCOM;
 using CommunityToolkit.WinUI.UI.Controls;
 using Hi3Helper;
 using Hi3Helper.Preset;
@@ -1346,7 +1347,7 @@ namespace CollapseLauncher.Pages
 
             LogWriteLine($"Media pack is not installed!\r\n\t" +
                         $"If you encounter the 'cry_ware_unity' error, run this script as an administrator:\r\n\t\t" +
-                        $"{Path.Combine(AppFolder, "Misc", "InstallMediaPack.cmd")}",LogType.Warning, true);
+                        $"{Path.Combine(AppFolder, "Misc", "InstallMediaPack.cmd")}", LogType.Warning, true);
 
             // Skip dialog if user asked before
             if (GetAppConfigValue("HI3IgnoreMediaPack").ToBool())
@@ -1463,7 +1464,7 @@ namespace CollapseLauncher.Pages
 
             WatchOutputLog.Cancel();
         }
-#endregion
+        #endregion
 
         #region Open Button Method
         private void OpenGameFolderButton_Click(object sender, RoutedEventArgs e)
